#173 move last of real duty functions into FileHub class
