+++ src/real/admin_duty.py
@@ -1,25 +1,9 @@
-from src._road.jaar_config import init_change_id
 from src._road.road import RoadUnit
 from src.agenda.group import GroupID
 from src.agenda.agenda import AgendaUnit
 from src.agenda.pledge import create_pledge
 from src.change.filehub import filehub_shop, FileHub
 from copy import deepcopy as copy_deepcopy
-
-
-class Invalid_duty_Exception(Exception):
-    pass
-
-
-def initialize_change_duty_files(x_filehub: FileHub):
-    x_duty_file_exists = x_filehub.duty_file_exists()
-    change_file_exists = x_filehub.change_file_exists(init_change_id())
-    if x_duty_file_exists == False and change_file_exists == False:
-        x_filehub._create_initial_change_and_duty_files()
-    elif x_duty_file_exists == False and change_file_exists:
-        x_filehub._create_duty_from_changes()
-    elif x_duty_file_exists and change_file_exists == False:
-        x_filehub._create_initial_change_files_from_duty()
 
 
 def append_changes_to_duty_file(x_filehub: FileHub) -> AgendaUnit:
